requre chalk in another try-catch
in the use case when eslint is installed locally with 3rd party
configurations, for example: eslint-config-airbnb. However chalk is not
installed locally, this will cause an exception and goes into the catch
clause and load eslint and chak from eslint_d.

Now eslint will not able to find eslint-config-airbnb which installed
locally.

I split the try-catch clause, so that eslint and chalk will be loaded
separately. Means eslint will be resolved from local, and whether there
is eslint at local, it does not effect the way how chalk will be loaded.