Check mtime of common package manager files
On every invocation, check the modified timestamp of the common package
manager files `package.json`, `package-lock.json` and `yarn.lock`. The
newest timestamp is compared with the last eslint execution time. If a
modification occurred in the meantime, the cached instance of eslint is
discarded and a new one is created.