Fix issue #101
browserify/resolve doesn't fully comply with node resolve algorithm. Change browserify/resolve to node's require.resolve.
require.resolve is injected in each module and cannot be spied by sinon in the same way how it works with browserify/resolve,
which is just a module.
Implement require.resolve inside resolver module in order to be able to spy on it in tests.